240104 boards

+++ SpringBoard-mybatis/src/main/webapp/WEB-INF/views/user/userlist.jsp
@@ -11,6 +11,12 @@
 <script src="https://kit.fontawesome.com/69798321c6.js" ></script>
 </head>
 <body>
+	<c:if test="${empty sessionId }">
+		<script type="text/javascript">
+			alert("로그인 후 이용 가능합니다.");
+			location.href = "/user/login";
+		</script>
+	</c:if>
 	<jsp:include page="../layout/header.jsp"/>
 	<div id="content">
 		<h2>회원 목록</h2>
@@ -33,7 +39,7 @@
 				<td>${user.userPasswd }</td>
 				<td>${user.userName }</td>
 				<td>${user.userAge }</td>
-				<td><a href="/user/delete?id=${user.id }"><i class="fa-solid fa-square-xmark"></i></a></td>
+				<td><a href="/user/delete?id=${user.id }" onclick="return confirm('정말로 삭제하시겠습니까?')"><i class="fa-solid fa-square-xmark"></i></a></td>
 			</tr>
 			</tbody>
 			</c:forEach>
